+++ udata_front/theme/gouvfr/assets/js/admin.ts
@@ -17,6 +17,7 @@ import Me from "./pages/Admin/Me/Me.vue";
 import Admin from "./pages/Admin/Admin.vue";
 import Members from "./pages/Admin/Members/Members.vue";
 import Profile from "./pages/Admin/Profile/Profile.vue";
+import Discussions from "./pages/Admin/Discussions/Discussions.vue";
 
 setupComponents({
   admin_root,
@@ -37,6 +38,7 @@ const routes: Array<RouteRecordRaw> = [
     children: [
       { path: 'datasets', component: Datasets, props: true, name: "organization-datasets" },
       { path: 'reuses', component: Reuses, props: true, name: "organization-reuses" },
+      { path: 'discussions', component: Discussions, props: true, name: "organization-discussions" },
       { path: 'members', component: Members, props: true, name: "organization-members" },
       { path: 'profile', component: Profile, props: true, name: "organization-profile" },
     ],