+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Adapt to discussion csv export refactor [#543](https://github.com/datagouv/udata-front/pull/543)
+- Add beta admin discussion page [#539](https://github.com/datagouv/udata-front/pull/539)
 - Fix Beta admin profile page [#546](https://github.com/datagouv/udata-front/pull/546)
 
 ## 5.2.2 (2024-09-23)